#15088 remove locked layout count

+++ src/server-base/src/main/java/io/kyligence/kap/rest/response/NDataModelResponse.java
@@ -75,7 +75,7 @@ public class NDataModelResponse extends NDataModel {
     private int recommendationsCount;
 
     @JsonProperty("available_indexes_count")
-    private int availableIndexesCount;
+    private long availableIndexesCount;
 
     private long lastModify;
